prototype/pyproject.toml
[tool.poetry]
name = "digital-emblem"
version = "0.2.0"
description = ""
authors = [
    "Nivesh Varma <nivesh.varma@jhuapl.edu>",
    "Rahel Fainchtein <rahel.fainchtein@jhuapl.edu>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cryptography = "~43.0.1"
plac = "^1.4.0"
acme = "^2.11.0"
dnspython = "^2.7.0"
certbot = "^2.11.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "fpho"
url = "https://files.pythonhosted.org"
priority = "primary"

[tool.poetry.group.dev.dependencies]
jedi = "^0.19.1"
ipython = "^8.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
